Update README.md with corrent repo
"feross" -> "standard". All updated static content links verified to be valid. But feel free to verify for yourself, too.

I wasn't positive on the license line if it should be changed to a standardjs.com name or not-- someone else can comment about that if it should also be updated.